@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
 
@@ -10,12 +10,15 @@
 
 
   <ItemGroup>
-    <PackageReference Include="Humanizer" Version="2.6.2" />
-    <PackageReference Include="AsyncEnumerator" Version="2.2.2" />
+    <PackageReference Include="Humanizer" Version="2.8.11" />
+    <PackageReference Include="AsyncEnumerator" Version="4.0.2" />
     <!-- <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-beta" /> -->
-    <PackageReference Include="Dapper" Version="1.60.6" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.0.0-preview6.19304.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.0.0-preview6.19304.10" />
+    <PackageReference Include="Dapper" Version="2.0.35" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.3" />
+
   </ItemGroup>
 
 </Project>
